requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
bidict==0.23.1
causal-conv1d==1.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
chex==0.1.88
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
diffusers==0.32.2
dill==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
dynamax==0.1.5
einops==0.8.0
einx==0.3.0
etils==1.11.0
exceptiongroup==1.2.2
executing==2.1.0
fastprogress==1.0.3
filelock==3.16.1
fonttools==4.55.4
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
huggingface-hub==0.27.1
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
jax==0.4.33
jax-cuda12-pjrt==0.4.33
jax-cuda12-plugin==0.4.33
jaxlib==0.4.33
jaxtyping==0.2.36
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lightning==2.5.0.post0
lightning-utilities==0.11.9
loguru==0.7.3
mamba-ssm==2.2.4
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mizani==0.13.1
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nnsight==0.3.7
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.6.85
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opt_einsum==3.4.0
optax==0.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotnine==0.14.5
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytorch-lightning==2.5.0.post0
pytz==2024.2
pyvene==0.1.6
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.12.0
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
statsmodels==0.14.4
submitit==1.5.2
sympy==1.13.1
tensorflow-probability==0.25.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.4.0
torch2jax==0.0.1
torchmetrics==1.6.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wandb==0.19.4
wcwidth==0.2.13
websocket-client==1.8.0
widgetsnbextension==4.0.13
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
